+++ src/test/extension.test.ts
@@ -10,7 +10,8 @@ suite("sortLines()", () => {
     async function doTest(
         input: string,
         selections: Selection[],
-        sortDescending: boolean) {
+        descending: boolean,
+        mode: my.Mode) {
 
         // Set the test input text and the selection
         const editor = vscode.window.activeTextEditor!;
@@ -26,7 +27,7 @@ suite("sortLines()", () => {
         editor.selections = selections;
 
         // Call the logic
-        await my.sortLines(editor, sortDescending);
+        await my.sortLines(editor, descending, mode);
 
         // Return the result
         return { text: editor.document.getText(), selections: editor.selections };
@@ -64,7 +65,7 @@ suite("sortLines()", () => {
             "c\n" +
             "b\n",
             [new Selection(0, 0, 3, 0)],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "a\n" +
@@ -80,7 +81,7 @@ suite("sortLines()", () => {
             "aa\n" +
             "a\n",
             [new Selection(0, 0, 3, 0)],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "a\n" +
@@ -98,7 +99,7 @@ suite("sortLines()", () => {
             "Orange\n" +
             "Grape\n",
             [new Selection(0, 3, 2, 4)],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "Apple\n" +
@@ -120,7 +121,7 @@ suite("sortLines()", () => {
                 new Selection(1, 1, 1, 2),
                 new Selection(2, 1, 2, 2),
             ],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "Pineapple\n" +
@@ -147,7 +148,7 @@ suite("sortLines()", () => {
                 new Selection(1, 3, 1, 1),
                 new Selection(3, 1, 3, 3),
             ],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "Pineapple\n" +
@@ -170,7 +171,7 @@ suite("sortLines()", () => {
             "c\n" +
             "b\n",
             [new Selection(0, 0, 3, 0)],
-            true
+            true, ""
         );
         assert.equal(result.text,
             "c\n" +
@@ -186,7 +187,7 @@ suite("sortLines()", () => {
             "aa\n" +
             "aaa\n",
             [new Selection(0, 0, 3, 0)],
-            true
+            true, ""
         );
         assert.equal(result.text,
             "aaa\n" +
@@ -205,7 +206,7 @@ suite("sortLines()", () => {
             "b\n" +
             "\n",
             [new Selection(0, 0, 3, 0)],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "a\n" +
@@ -229,7 +230,7 @@ suite("sortLines()", () => {
                 new Selection(1, 0, 1, 0),
                 new Selection(2, 0, 2, 1),
             ],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "\n" +
@@ -267,11 +268,33 @@ suite("sortLines()", () => {
         let result = await doTest(
             input.join("\n") + "\n",
             [new Selection(0, 0, N * 2, 0)],
-            false
+            false, ""
         );
         assert.equal(result.text,
             expected.join("\n") + "\n"
         );
+    });
+
+    // Collate options
+    [
+        { descending: false, mode: "", expected: "10,2,２,ab,Ac,あ,ア" },
+        { descending: false, mode: "n", expected: "2,２,10,ab,Ac,あ,ア" },
+        { descending: false, mode: "c", expected: "10,2,Ac,ab,あ,ア,２" },
+        { descending: true, mode: "", expected: "あ,ア,Ac,ab,2,２,10" },
+        { descending: true, mode: "n", expected: "あ,ア,Ac,ab,10,2,２" },
+        { descending: true, mode: "c", expected: "２,ア,あ,ab,Ac,2,10" },
+    ].forEach(t => {
+        const input = "2,10,あ,ab,２,Ac,ア";
+        test(`options: {descending: ${t.descending}, mode: "${t.mode}"}`,
+            async () => {
+                const result = await doTest(
+                    input.replace(/,/g, "\n"),
+                    [new Selection(0, 0, 6, 1)],
+                    t.descending,
+                    t.mode as my.Mode
+                );
+                assert.equal(result.text.replace(/\n/g, ","), t.expected);
+            });
     });
 });
 
@@ -280,7 +303,8 @@ suite("sortWords()", () => {
     async function doTest(
         input: string,
         selections: Selection[],
-        sortDescending: boolean) {
+        descending: boolean,
+        mode: my.Mode) {
 
         // Set the test input text and the selection
         const editor = vscode.window.activeTextEditor!;
@@ -296,7 +320,7 @@ suite("sortWords()", () => {
         editor.selections = selections;
 
         // Call the logic
-        await my.sortWords(editor, sortDescending);
+        await my.sortWords(editor, descending, mode);
 
         // Return the result
         return { text: editor.document.getText(), selections: editor.selections };
@@ -320,7 +344,7 @@ suite("sortWords()", () => {
         let result = await doTest(
             "1, 2,10",
             [new Selection(0, 0, 0, 7)],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "1, 10, 2"
@@ -334,7 +358,7 @@ suite("sortWords()", () => {
         let result = await doTest(
             "1 2 10",
             [new Selection(0, 0, 0, 6)],
-            false
+            false, ""
         );
         assert.equal(result.text,
             "1 10 2"
@@ -342,6 +366,28 @@ suite("sortWords()", () => {
         assert.equal(stringifySelections(result.selections),
             stringifySelections([new Selection(0, 0, 0, 6)])
         );
+    });
+
+    // Collate options
+    [
+        { descending: false, mode: "", expected: "10,2,２,ab,Ac,あ,ア" },
+        { descending: false, mode: "n", expected: "2,２,10,ab,Ac,あ,ア" },
+        { descending: false, mode: "c", expected: "10,2,Ac,ab,あ,ア,２" },
+        { descending: true, mode: "", expected: "あ,ア,Ac,ab,2,２,10" },
+        { descending: true, mode: "n", expected: "あ,ア,Ac,ab,10,2,２" },
+        { descending: true, mode: "c", expected: "２,ア,あ,ab,Ac,2,10" },
+    ].forEach(t => {
+        const input = "2,10,あ,ab,２,Ac,ア";
+        test(`options: {descending: ${t.descending}, mode: "${t.mode}"}`,
+            async () => {
+                const result = await doTest(
+                    input,
+                    [new Selection(0, 0, 0, 16)],
+                    t.descending,
+                    t.mode as my.Mode
+                );
+                assert.equal(result.text, t.expected);
+            });
     });
 });